feat: dynamically load debt calendar
